+++ CHANGES.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+- HOLLight_Real.Quotient: remove Local Definition of a
+- hol_upto_real.ml: update wrt HOL-Light 3.0
+- erasing.lp: update wrt https://github.com/Deducteam/hol2dk/pull/161 and https://github.com/Deducteam/hol2dk/pull/149
+- reproduce: update hollight, hol2dk and lambdapi versions
+- reproduce: add stage mechanism to run reproduce again by skipping stages that succeeded
+- terms.v, theorems.v: type arguments have readable names now
+
 ## 1.0.0 (2024-11-03)
 
 First release.